It has been compiled successfully, but there is an issue with the data flow. Sending a Resize to vim failed.
