+++ src/messager.rs
@@ -8,48 +8,62 @@ use crate::{
     running_tracker::RUNNING_TRACKER,
 };
 
+#[derive(Debug)]
 pub struct VimMessager {}
 
-pub struct RuntimeRef<'a>(&'a tokio::runtime::Runtime);
-
-impl Worker for VimMessager {
-    type Init = RuntimeRef<'_>;
+impl Component for VimMessager {
+    type Init = ();
     type Input = RedrawEvent;
-    type Output = AppMessage;
+    type Output = ();
+    type Root = ();
+    type Widgets = ();
+    type CommandOutput = AppMessage;
 
-    fn init(RuntimeRef(rt): Self::Init, sender: ComponentSender<Self>) -> Self {
-        let mut rx = EVENT_AGGREGATOR.register_event::<RedrawEvent>();
-        let sender = sender.clone();
-        let running_tracker = RUNNING_TRACKER.clone();
-        rt.spawn(async move {
-            loop {
-                tokio::select! {
-                    _ = running_tracker.wait_quit() => {
-                        info!("messager quit.");
-                        sender.output(AppMessage::Quit).unwrap();
-                        // 保证最后一个退出, 避免其他task还在写,这里已经关闭,报错.
-                        tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-                        break;
-                    },
-                    Some(event) = rx.recv() => {
-                        trace!("RedrawEvent {:?}", event);
-                        sender
-                            .output(AppMessage::RedrawEvent(event))
-                            .expect("Failed to send RedrawEvent to main thread");
-                    },
-                    else => {
-                        info!("messager None RedrawEvent event received, quit.");
-                        sender.output(AppMessage::Quit).unwrap();
-                        break;
-                    },
-                }
-            }
-        });
-
-        VimMessager {}
+    fn init_root() -> Self::Root {
+        ()
     }
 
-    fn update(&mut self, message: RedrawEvent, _: ComponentSender<Self>) {
+    fn init(_: Self::Init, _: &Self::Root, sender: ComponentSender<Self>) -> ComponentParts<Self> {
+        let mut rx = EVENT_AGGREGATOR.register_event::<RedrawEvent>();
+        // let sender = sender.clone();
+        let running_tracker = RUNNING_TRACKER.clone();
+
+        sender.command(|out, shutdown| {
+            shutdown
+                .register(async move {
+                    loop {
+                        tokio::select! {
+                            _ = running_tracker.wait_quit() => {
+                                info!("messager quit.");
+                                out.send(AppMessage::Quit).unwrap();
+                                // 保证最后一个退出, 避免其他task还在写,这里已经关闭,报错.
+                                tokio::time::sleep(std::time::Duration::from_millis(10)).await;
+                                break;
+                            },
+                            Some(event) = rx.recv() => {
+                                trace!("RedrawEvent {:?}", event);
+                                out
+                                    .send(AppMessage::RedrawEvent(event))
+                                    .expect("Failed to send RedrawEvent to main thread");
+                            },
+                            else => {
+                                info!("messager None RedrawEvent event received, quit.");
+                                out.send(AppMessage::Quit).unwrap();
+                                break;
+                            },
+                        }
+                    }
+                })
+                .drop_on_shutdown()
+        });
+
+        ComponentParts {
+            model: VimMessager {},
+            widgets: (),
+        }
+    }
+
+    fn update(&mut self, message: Self::Input, _sender: ComponentSender<Self>, _: &Self::Root) {
         EVENT_AGGREGATOR.send::<RedrawEvent>(message);
     }
 }
